+++ lib/modules/perf/perf_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: perf
-// Updated at: 2026-08-10 19:41:26
-// Commit: feat(perf): implement responsive perf view component
+// Updated at: 2026-08-11 15:06:07
+// Commit: test(perf): implement widget integration tests for perf
 
 class perfComponentState {
-  final String commitRef = '3505';
-  final String lastMessage = 'feat(perf): implement responsive perf view component';
-  final DateTime timestamp = DateTime.parse('2026-08-10T19:41:26+05:00');
+  final String commitRef = '3529';
+  final String lastMessage = 'test(perf): implement widget integration tests for perf';
+  final DateTime timestamp = DateTime.parse('2026-08-11T15:06:07+05:00');
 }